feat: reactive system
